# Conflicts:
#	app/assets/styles/app.css
#	app/templates/absences/index.html.twig
#	app/templates/dashboard/index.html.twig
#	app/templates/documents/index.html.twig
#	app/templates/grades/index.html.twig
#	app/templates/grades/show.html.twig
#	app/templates/projects/index.html.twig
#	app/templates/projects/show.html.twig
#	app/templates/promotion_users/index.html.twig
#	app/templates/promotions/index.html.twig

+++ app/src/Controller/PromotionsController.php
@@ -6,6 +6,7 @@ use App\Entity\Promotions;
 use App\Form\PromotionsType;
 use App\Repository\PromotionsRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -17,10 +18,22 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 final class PromotionsController extends AbstractController
 {
     #[Route(name: 'app_promotions_index', methods: ['GET'])]
-    public function index(PromotionsRepository $promotionsRepository): Response
+    public function index(PromotionsRepository $promotionsRepository, PaginatorInterface $paginator, Request $request): Response
     {
+        $user = $this->getUser();
+
+        if ($this->isGranted('ROLE_ADMIN') || $this->isGranted('ROLE_VISITOR')) {
+            $data = $promotionsRepository->findAll();
+        } elseif ($this->isGranted('ROLE_TEACHER')) {
+            $data = $promotionsRepository->findByTeacher($user);
+        } else {
+            $data = $promotionsRepository->findByStudent($user);
+        }
+
+        $promotions = $paginator->paginate($data, $request->query->getInt('page', 1), 20);
+
         return $this->render('promotions/index.html.twig', [
-            'promotions' => $promotionsRepository->findAll(),
+            'promotions' => $promotions,
         ]);
     }
 
@@ -31,6 +44,8 @@ final class PromotionsController extends AbstractController
         $promotion = new Promotions();
         $form = $this->createForm(PromotionsType::class, $promotion);
         $form->handleRequest($request);
+
+
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($promotion);
@@ -57,6 +72,10 @@ final class PromotionsController extends AbstractController
     #[IsGranted('ROLE_TEACHER')]
     public function edit(Request $request, Promotions $promotion, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') && $promotion->getProfessor()?->getId() !== $this->getUser()->getId()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez modifier que vos propres promotions.');
+        }
+
         $form = $this->createForm(PromotionsType::class, $promotion);
         $form->handleRequest($request);
 
@@ -76,7 +95,11 @@ final class PromotionsController extends AbstractController
     #[IsGranted('ROLE_TEACHER')]
     public function delete(Request $request, Promotions $promotion, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$promotion->getId(), $request->getPayload()->getString('_token'))) {
+        if (!$this->isGranted('ROLE_ADMIN') && $promotion->getProfessor()?->getId() !== $this->getUser()->getId()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez supprimer que vos propres promotions.');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $promotion->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($promotion);
             $entityManager->flush();
         }
